common base class for ModButton and ModLayoutButton
