+++ OWML.ModHelper.Menus/OWML.ModHelper.Menus.csproj
@@ -75,6 +75,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="BaseButton.cs" />
     <Compile Include="ModLayoutButton.cs" />
     <Compile Include="ModButton.cs" />
     <Compile Include="ModConfigMenu.cs" />
